+++ host/include/uhd/usrp/single_usrp.hpp
@@ -190,10 +190,10 @@ public:
      * \param name the name of the gain element
      * \param chan the channel index 0 to N-1
      */
-    virtual void set_rx_gain(float gain, const std::string &name, size_t chan = 0) = 0;
+    virtual void set_rx_gain(double gain, const std::string &name, size_t chan = 0) = 0;
 
     //! A convenience wrapper for setting overall RX gain
-    void set_rx_gain(float gain, size_t chan = 0){
+    void set_rx_gain(double gain, size_t chan = 0){
         return this->set_rx_gain(gain, ALL_GAINS, chan);
     }
 
@@ -204,10 +204,10 @@ public:
      * \param chan the channel index 0 to N-1
      * \return the gain in dB
      */
-    virtual float get_rx_gain(const std::string &name, size_t chan = 0) = 0;
+    virtual double get_rx_gain(const std::string &name, size_t chan = 0) = 0;
 
     //! A convenience wrapper for getting overall RX gain
-    float get_rx_gain(size_t chan = 0){
+    double get_rx_gain(size_t chan = 0){
         return this->get_rx_gain(ALL_GAINS, chan);
     }
 
@@ -281,7 +281,7 @@ public:
      * \return the rssi in dB
      * \throw exception if RSSI readback not supported
      */
-    virtual float read_rssi(size_t chan = 0) = 0;
+    virtual double read_rssi(size_t chan = 0) = 0;
 
     /*!
      * Get the dboard interface object for the RX subdevice.
@@ -359,10 +359,10 @@ public:
      * \param name the name of the gain element
      * \param chan the channel index 0 to N-1
      */
-    virtual void set_tx_gain(float gain, const std::string &name, size_t chan = 0) = 0;
+    virtual void set_tx_gain(double gain, const std::string &name, size_t chan = 0) = 0;
 
     //! A convenience wrapper for setting overall TX gain
-    void set_tx_gain(float gain, size_t chan = 0){
+    void set_tx_gain(double gain, size_t chan = 0){
         return this->set_tx_gain(gain, ALL_GAINS, chan);
     }
 
@@ -373,10 +373,10 @@ public:
      * \param chan the channel index 0 to N-1
      * \return the gain in dB
      */
-    virtual float get_tx_gain(const std::string &name, size_t chan = 0) = 0;
+    virtual double get_tx_gain(const std::string &name, size_t chan = 0) = 0;
 
     //! A convenience wrapper for getting overall TX gain
-    float get_tx_gain(size_t chan = 0){
+    double get_tx_gain(size_t chan = 0){
         return this->get_tx_gain(ALL_GAINS, chan);
     }
